requirements.txt
django
#django-countries
pillow
djangorestframework
drf-yasg
psycopg2-binary
